requirements.txt
fastapi
uvicorn
motor               # ✅ Async MongoDB client (replaces pymongo)
httpx               # ✅ Async HTTP client (replaces requests)

python-dotenv
openai
aiohttp             # ✅ Required for openai>=1.0.0+
fuzzywuzzy
python-Levenshtein>=0.12.0

Jinja2
python-multipart
pymongo
itsdangerous
pydantic>=2.0
pydantic_ai>=0.1.5
typing_extensions>=4.5


# === Core ===
pydantic>=2.4
pydantic-ai>=0.3.0
openai>=1.30.1

# === Scrapers ===
beautifulsoup4
selenium
requests
yt-dlp
youtube-search-python
huggingface_hub

# === MongoDB ===
pymongo
python-dotenv

# === Misc ===
tqdm
fastapi 
uvicorn 
motor
certifi


pymupdf 
pdf2image 
requests 
pymongo


python-telegram-bot
nest_asyncio 

markdown
fpdf


httpx>=0.27.0

openpyxl


pytesseract 
pdf2image 
pillow

